Add separate Anthropic proxy route for preview.new
Extract shared proxy logic into createAnthropicProxyHandler factory.
- /api/anthropic/v1/messages uses cmux-ai-proxy gateway
- /api/preview/anthropic/v1/messages uses cmux-preview-new gateway
