+++ apps/www/app/api/preview/anthropic/v1/messages/route.ts
@@ -1,0 +1,9 @@
+import { createAnthropicProxyHandler } from "../../../../anthropic/proxy-handler";
+
+const ANTHROPIC_API_URL =
+  "https://gateway.ai.cloudflare.com/v1/0c1675e0def6de1ab3a50a4e17dc5656/cmux-preview-new/anthropic/v1/messages";
+
+export const POST = createAnthropicProxyHandler(
+  ANTHROPIC_API_URL,
+  "anthropic preview proxy"
+);
